+++ src/tools/webhooks.ts
@@ -1,3 +1,4 @@
+import { isNode, isReactNative } from '../constants';
 import { SearchResult, Webhook, ApiEvent, ApiEventType } from '../types';
 import { WrapperClient } from '../wrapper';
 
@@ -65,11 +66,67 @@ export default class Webhooks {
    * @param payload - Received event object to validate
    * @returns When the signature is valid, it returns the event object
    */
-  async validateSignature<T extends ApiEventType | '' = ''>(data: {
+  async validateSignature<T extends ApiEventType>(data: {
     secret: string;
     signature: string;
-    payload: ApiEvent<T>;
+    payload: string | Buffer | ApiEvent<T>;
   }): Promise<ApiEvent<T>> {
-    return this.client.post('/webhooks/validate-signature', { body: data });
+    // Validated locally
+    const { secret, signature, payload } = data;
+    let payloadString: string;
+    if (typeof payload === 'string') {
+      payloadString = payload;
+    } else if (Buffer.isBuffer(payload)) {
+      payloadString = payload.toString('utf8');
+    } else if (typeof payload === 'object') {
+      payloadString = JSON.stringify(payload);
+    } else {
+      throw new Error('Invalid payload type');
+    }
+
+    if (isNode) {
+      const crypto = await import('crypto');
+      const hmac = crypto.createHmac('sha256', secret);
+      const digest = hmac
+        .update(payloadString)
+        .digest('hex')
+        .toLowerCase();
+      if (signature !== digest) {
+        throw new Error('Invalid signature');
+      }
+      return JSON.parse(payloadString) as ApiEvent<T>;
+    } else if (isReactNative) {
+      // Call the API to validate signature in React Native
+      return this.client.post('/webhooks/validate-signature', {
+        body: {
+          secret,
+          signature,
+          payload: payloadString,
+        },
+      });
+    } else { // Web browsers
+      const encoder = new TextEncoder();
+      const encodedData = encoder.encode(payloadString);
+      const encodedSecret = encoder.encode(secret);
+      const digest = await crypto.subtle.sign(
+        'HMAC',
+        await crypto.subtle.importKey(
+          'raw',
+          encodedSecret,
+          { name: 'HMAC', hash: 'SHA-256' },
+          false,
+          ['sign'],
+        ),
+        encodedData,
+      );
+      const hexDigest = Array.from(new Uint8Array(digest))
+        .map((b) => b.toString(16).padStart(2, '0'))
+        .join('')
+        .toLowerCase();
+      if (signature !== hexDigest) {
+        throw new Error('Invalid signature');
+      }
+      return JSON.parse(payloadString) as ApiEvent<T>;
+    }
   }
 }