+++ vite.config.ts
@@ -8,7 +8,7 @@ export default defineConfig({
       fileName: (format) => `index.${format}.js`,
     },
     rollupOptions: {
-      external: ['stream'],
+      external: ['stream', 'crypto'],
       output: {
         exports: 'named', // Use named exports to avoid the warning
       },